requirements.txt
Brotli 
certifi 
charset-normalizer 
idna 
jplephem==2.22
numpy==2.1.3
PySocks 
python-dateutil==2.9.0.post0
pytz==2024.2
requests 
sgp4==2.23
six==1.16.0
skyfield==1.49
urllib3 
win-inet-pton 
